Feat: 유저 입장 로직 변경
